рефакторит создание баллунов
